rfct!: More readable and sensible normal user list
Improve the readablity of the process of loading user and default shell
info from the `/etc/passwd` file.

BREAKING CHANGE: Flip the binding on the users struct. Before it was
fancy name -> system username, now it is: system -> fancy. The previous
order is super unexpected. Also can cause issues if 2 users have the
same full name in the gecos field.