ui: Remove the reconnect button (not planned)
The reconnect button was intended as a way to recover from weird IPC
states. However greetd doesn't work like other software where each
connection has it's own state so reconnecting to it does not trigger a
reset. Thus we just have to be careful and issue cancel requests when
something goes wrong.