rfct: Resolve potential panic cond. in created session state
The username cannot be changed once authentication was started with
greetd. However it is theoretically possible to change the username in
the ui before the input receives the lock event.

To resolve this the input is hidden and replaced with an immutable
label that displays the state of the current username once the session
is created for that user. Once the session is canceled, the input is
told to sync up with the previous user value.

This way the UI is guaranteed to be consistent with the IPC state and
the messages to change user can be safely ignored even when they should
not be emited.